reservation components
