+++ app/src/main/java/reservant_mobile/ui/activities/RestaurantDetailActivity.kt
@@ -58,7 +58,7 @@ import kotlinx.coroutines.launch
 import reservant_mobile.data.constants.Roles
 import reservant_mobile.data.models.dtos.RestaurantMenuDTO
 import reservant_mobile.data.models.dtos.RestaurantMenuItemDTO
-import reservant_mobile.ui.components.FloatingActionMenu
+import reservant_mobile.ui.components.ReservationFloatingMenu
 import reservant_mobile.ui.components.FloatingTabSwitch
 import reservant_mobile.ui.components.FullscreenGallery
 import reservant_mobile.ui.components.ImageCard
@@ -264,7 +264,7 @@ fun RestaurantDetailActivity(navController: NavController, restaurantId: Int) {
             }
         }
     }
-    FloatingActionMenu(
+    ReservationFloatingMenu(
         onDineInClick = {},
         onDeliveryClick = {},
         onTakeawayClick= {}
